requirements.txt
argparse==1.4.0
matplotlib==3.9.0
numpy==1.26.4
open3d==0.18.0
opencv-python==4.6.0.66
pyyaml===6.0.2
scipy==1.13.1
tqdm==4.66.5
ninja==1.11.1.1